pyproject.toml
[project]
name = "constrained-toasts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.5",
]
